requirements.txt
jupyter==1.0.0
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
scikit-learn==0.22.2.post1
tensorboard==2.2.1
torch==1.4.0
torchvision==0.5.0
